chore: start ElasticSearchQuerier
